requirements.txt
colorcet
pynwb
Pillow
tqdm
PySide2==5.15.2 
pyqt5
pyqtwebengine
pyqt5-sip
pyqtgraph==0.12.4
PyOpenGL==3.1.6
PyOpenGL_accelerate==3.1.6
cython
neo
pylru==1.2.1
h5py==3.7.0
joblib
future==0.18.2
pyYAML==6.0
python-dateutil
pytz
scipy
pandas
matplotlib
scikit-image